requirements.txt
asgiref==3.5.2
Django==3.2.16
django-filter==22.1
django-jazzmin==2.6.0
django-storages==1.13.1
djangorestframework==3.14.0
importlib-metadata==5.0.0
Markdown==3.4.1
pytz==2022.6
sqlparse==0.4.3
zipp==3.10.0
psycopg2-binary==2.9.5
django-redis==5.2.0
gunicorn==20.1.0
Pillow==9.2.0
uwsgi>=2.0.18,<2.1.0
s3transfer==0.5.2
boto3==1.20.44
botocore==1.23.44
django-filter